Update navigation.yaml to match the default for nav2 on jazzy
The navigation.yaml in the repo almost completely dates back to pre-jazzy, and its layout
was so different from the default that nav2 on Jazzy deposits in /tmp when run with:
ros2 launch nav2_bringup bringup_launch.py use_sim_time:=False autostart:=True map:=/path/to/your-map.yaml
that comparison wasn't really possible. Furthermore, the nav stack was unable to navigate to a goal
using the previous navigation.yaml.

Problems found were:
1. it had a syntax error in naming a plugin:
line 368 contained: plugin: "nav2_navfn_planner/NavfnPlanner"
should have been: plugin: "nav2_navfn_planner::NavfnPlanner"
so it could never work.
2. After fixing this, attempting to navigate gave errors like:
[component_container_isolated-1] [INFO] [1741063958.773577902] [amcl]: Setting pose (1741063958.773578): -0.104 -0.005 1.851 [rviz2-2] Start navigation [rviz2-2] [ERROR] [1741063991.586034117] [rviz_navigation_dialog_action_client]: navigate_to_pose action server is not available. Is the initial pose set?

The best course of action seems to be to bring it up to date with the default Jazzy nav2 config,
which does navigate a real 2wd robot successfully.

Thomas Chou had made 2 changes to navigation.yaml:
1. Reduced the 3rd parameter in max and min speed from 2.5 to 2.0. This seems like a difference
not appropriate for incorporating in a "basic-working" navigation.yaml
2. Added sonar support in commit on 2024-07-06. I looked at adding it into the nav2 default which
is the subject of this PR, but the format of configuration around it was so different that it
couldn't be put in the same way it was in the previous file. For example, the prevous file had
the global planner looking at the voxel layer, and T. Chou had added a sonar layer to the
global planner, whereas the current jazzy default does not look at the voxel layer in the
global planner, only in the local planner. Also, the format for declaring plugins seemed to
have changed. I recommend that if a sonar layer should be re-introduced,
it is done in the context of this Jazzy nav2 config.